Probe LA independent of housekeeping interval
Load Average is a metric that inherently relies on being probed for
frequently (otherwise, it ends up being dependent on the luck of the
draw more so than on the actual value). This updates the container
manager to run the LA probe on a different schedule from the regular
housekeeping, and uses the last observed value when reporting.